pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-sales-analysis"
version = "1.0.0"
description = "Multi-agent sales analysis system with LLM + function hybrid approach"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["sales", "analysis", "llm", "multi-agent", "ai", "data-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.0.20",
    "google-generativeai>=0.3.2",
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "openpyxl>=3.1.2",
    "xlrd>=2.0.1",
    "google-auth>=2.25.2",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.111.0",
    "boto3>=1.34.26",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.0",
    "scipy>=1.11.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
sales-agent = "sales_agent.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/multi-agent-sales-analysis"
Documentation = "https://github.com/yourusername/multi-agent-sales-analysis#readme"
Repository = "https://github.com/yourusername/multi-agent-sales-analysis"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sales_agent*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
